app/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
scikit-learn==1.7.2
nltk==3.9.2
textblob==0.19.0
pandas==2.3.3
numpy==2.3.4
python-dotenv==1.2.1
motor==3.3.1
pymongo==4.5.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
requests==2.32.5
PyPDF2==4.0.1
python-multipart==0.0.20
pydantic==2.12.3

